Add tests for Order API. Move catalog object creation to shared context
* Add Tests for Order API
* Move catalog objects creation to a shared context
