+++ spec/api/catalog_api_spec.rb
@@ -15,276 +15,14 @@ require 'securerandom'
 # Automatically generated by swagger-codegen (github.com/swagger-api/swagger-codegen)
 # Please update as you see appropriate
 describe 'CatalogApi' do
-
-  let(:beverages_client_id) { '#Beverages' }
-  let(:beverages) do
-    {
-      type: SquareConnect::CatalogObjectType::CATEGORY,
-      id: beverages_client_id,
-      category_data: {
-        name: 'Beverages'
-      }
-    }
-  end
-
-  let(:sales_tax_client_id) { '#SalesTax' }
-  let(:sales_tax) do
-    {
-      type: SquareConnect::CatalogObjectType::TAX,
-      id: sales_tax_client_id,
-      present_at_all_locations: true,
-      tax_data: {
-        name: 'Sales Tax',
-        calculation_phase: SquareConnect::TaxCalculationPhase::SUBTOTAL_PHASE,
-        inclusion_type: SquareConnect::TaxInclusionType::ADDITIVE,
-        percentage: '5.0',
-        applies_to_custom_amounts: true,
-        enabled: true
-      }
-    }
-  end
-
-  let(:milks_client_id) { '#Milks' }
-  let(:whole_milk_client_id) { '#WholeMilk' }
-  let(:skim_milk_client_id) { '#SkimMilk' }
-  let(:soy_milk_client_id) { '#SoymMilk' }
-  let(:milks) do
-    {
-      type: SquareConnect::CatalogObjectType::MODIFIER_LIST,
-      id: milks_client_id,
-      modifier_list_data: {
-        name: 'Milks',
-        modifiers: [
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: whole_milk_client_id,
-            modifier_data: {
-              name: 'Whole Milk'
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: skim_milk_client_id,
-            modifier_data: {
-              name: 'Skim Milk'
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: soy_milk_client_id,
-            modifier_data: {
-              name: 'Soy Milk',
-              price_money: {
-                amount: 50,
-                currency: SquareConnect::Currency::USD
-              }
-            }
-          },
-        ]
-     }
-    }
-  end
-
-  let(:syrups_client_id) { '#Syrups' }
-  let(:hazelnut_client_id) { '#Hazelnut' }
-  let(:vanilla_client_id) { '#Vanilla' }
-  let(:chocolate_client_id) { '#Chocolate' }
-  let(:syrups) do
-    {
-      type: SquareConnect::CatalogObjectType::MODIFIER_LIST,
-      id: syrups_client_id,
-      modifier_list_data: {
-        name: 'Syrups',
-        modifiers: [
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: hazelnut_client_id,
-            modifier_data: {
-              name: 'Hazelnut'
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: vanilla_client_id,
-            modifier_data: {
-              name: 'Vanilla'
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::MODIFIER,
-            id: chocolate_client_id,
-            modifier_data: {
-              name: 'Chocolate'
-            }
-          },
-        ]
-      }
-    }
-  end
-
-  let(:coffee_client_id) { '#Coffee' }
-  let(:small_coffee_client_id) { '#SmallCoffee' }
-  let(:large_coffee_client_id) { '#LargeCoffee' }
-  let(:coffee) do
-    {
-      type: SquareConnect::CatalogObjectType::ITEM,
-      id: coffee_client_id,
-      present_at_all_locations: true,
-      item_data: {
-        name: 'Coffee',
-        description: 'Hot bean juice',
-        abbreviation: 'Co',
-        category_id: beverages_client_id,
-        modifier_list_info: [
-          {
-            modifier_list_id: milks_client_id
-          }
-        ],
-        tax_ids: [sales_tax_client_id],
-        variations: [
-          {
-            type: SquareConnect::CatalogObjectType::ITEM_VARIATION,
-            id: small_coffee_client_id,
-            present_at_all_locations: true,
-            item_variation_data: {
-              item_id: coffee_client_id,
-              name: 'Small',
-              pricing_type: SquareConnect::CatalogPricingType::FIXED_PRICING,
-              price_money: {
-                amount: 195,
-                currency: SquareConnect::Currency::USD
-              }
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::ITEM_VARIATION,
-            id: large_coffee_client_id,
-            present_at_all_locations: true,
-            item_variation_data: {
-              item_id: coffee_client_id,
-              name: 'Large',
-              pricing_type: SquareConnect::CatalogPricingType::FIXED_PRICING,
-              price_money: {
-                amount: 250,
-                currency: SquareConnect::Currency::USD
-              }
-            }
-          }
-        ]
-      }
-    }
-  end
-
-  let(:tea_client_id) { '#Tea' }
-  let(:small_tea_client_id) { '#SmallTea' }
-  let(:large_tea_client_id) { '#LargeTea' }
-  let(:tea) do
-    {
-      type: SquareConnect::CatalogObjectType::ITEM,
-      id: tea_client_id,
-      present_at_all_locations: true,
-      item_data: {
-        name: 'Tea',
-        description: 'Hot leaf juice',
-        abbreviation: 'Te',
-        category_id: beverages_client_id,
-        tax_ids: [sales_tax_client_id],
-        variations: [
-          {
-            type: SquareConnect::CatalogObjectType::ITEM_VARIATION,
-            id: small_tea_client_id,
-            present_at_all_locations: true,
-            item_variation_data: {
-              item_id: tea_client_id,
-              name: 'Small',
-              pricing_type: SquareConnect::CatalogPricingType::FIXED_PRICING,
-              price_money: {
-                amount: 150,
-                currency: SquareConnect::Currency::USD
-              }
-            }
-          },
-          {
-            type: SquareConnect::CatalogObjectType::ITEM_VARIATION,
-            id: large_tea_client_id,
-            present_at_all_locations: true,
-            item_variation_data: {
-              item_id: tea_client_id,
-              name: 'Large',
-              pricing_type: SquareConnect::CatalogPricingType::FIXED_PRICING,
-              price_money: {
-                amount: 200,
-                currency: SquareConnect::Currency::USD
-              }
-            }
-          }
-        ]
-      }
-    }
-  end
-
-  let(:object_definitions) { [beverages, milks, syrups, coffee, tea, sales_tax] }
+  include_context 'Catalog definitions'
 
   let(:accounts) { load_accounts() }
   let(:account) { accounts['US-Prod'] }
-  let(:instance) do
-    configuration = SquareConnect::Configuration.new do |config|
-      config.access_token = account["access_token"]
-    end
-    client = SquareConnect::ApiClient.new(configuration)
 
-    SquareConnect::CatalogApi.new(client)
-  end
-
-  let!(:id_map) do
-    # Set a unique idempotency key for each request
-    body = SquareConnect::BatchUpsertCatalogObjectsRequest.new
-    body.idempotency_key = SecureRandom.uuid
-    body.batches = [{
-      objects: object_definitions
-    }]
-
-    # Batch upsert the objects
-    response = instance.batch_upsert_catalog_objects(body)
-
-    mapping = response.id_mappings.map do |id_mapping|
-      [id_mapping.client_object_id, id_mapping.object_id]
-    end
-
-    mapping.to_h
-  end
-
-  let(:coffee_id) { id_map[coffee_client_id] }
-  let(:small_coffee_id) { id_map[small_coffee_client_id] }
-  let(:large_coffee_id) { id_map[large_coffee_client_id] }
-  let(:small_tea_id) { id_map[small_tea_client_id] }
-  let(:sales_tax_id)  { id_map[sales_tax_client_id] }
-  let(:beverages_id)  { id_map[beverages_client_id] }
-  let(:milks_id)  { id_map[milks_client_id] }
-  let(:syrups_id) { id_map[syrups_client_id] }
-
-  after(:each) { delete_test_catalog }
-
-  def delete_test_catalog
-    opts = {}
-    objects_to_delete = []
-
-    begin
-      response = instance.list_catalog(opts)
-      objects_to_delete += response.objects.map(&:id)
-      opts[:cursor] = response.cursor
-    end while opts[:cursor]
-
-    objects_to_delete.each_slice(200) do |object_ids|
-      body = SquareConnect::BatchDeleteCatalogObjectsRequest.new
-      body.object_ids = object_ids
-      instance.batch_delete_catalog_objects(body)
-    end
-  end
-
-  describe 'instance' do
-    it 'should create an instance of CatalogApi' do
-      expect(instance).to be_instance_of(SquareConnect::CatalogApi)
+  describe 'catalog_instance' do
+    it 'should create an catalog_instance of CatalogApi' do
+      expect(catalog_instance).to be_instance_of(SquareConnect::CatalogApi)
     end
   end
 
@@ -292,7 +30,7 @@ describe 'CatalogApi' do
     it 'should delete a set of items based on the provided list of target IDs and return the deleted objects' do
       body = SquareConnect::BatchDeleteCatalogObjectsRequest.new
       body.object_ids = [coffee_id, small_tea_id]
-      response = instance.batch_delete_catalog_objects(body)
+      response = catalog_instance.batch_delete_catalog_objects(body)
 
       expect(response.deleted_object_ids.size).to eq(4)
       expect(response.deleted_object_ids).to include(coffee_id)
@@ -306,7 +44,7 @@ describe 'CatalogApi' do
     it 'should return a set of objects based on the provided ID.' do
       body = SquareConnect::BatchRetrieveCatalogObjectsRequest.new
       body.object_ids = [coffee_id, sales_tax_id]
-      response = instance.batch_retrieve_catalog_objects(body)
+      response = catalog_instance.batch_retrieve_catalog_objects(body)
 
       expect(response.objects.size).to eq(2)
 
@@ -405,7 +143,7 @@ describe 'CatalogApi' do
       body.idempotency_key = SecureRandom.uuid
       body.batches = batches
 
-      response = instance.batch_upsert_catalog_objects(body)
+      response = catalog_instance.batch_upsert_catalog_objects(body)
 
       expect(response.objects.size).to eq(num_objects)
     end
@@ -413,7 +151,7 @@ describe 'CatalogApi' do
 
   describe '#catalog_info' do
     it 'should return information about Square Catalog API' do
-      result = instance.catalog_info()
+      result = catalog_instance.catalog_info()
       expect(result.limits.batch_upsert_max_objects_per_batch).to eq(1000)
       expect(result.limits.batch_upsert_max_total_objects).to eq(10000)
       expect(result.limits.batch_retrieve_max_object_ids).to eq(1000)
@@ -430,7 +168,7 @@ describe 'CatalogApi' do
 
   describe '#delete_catalog_object' do
     it 'should delete a single catalog object based on the provided ID and return a set of the successfully deleted object ids' do
-      response = instance.delete_catalog_object(coffee_id)
+      response = catalog_instance.delete_catalog_object(coffee_id)
 
       expect(response.deleted_object_ids.size).to eq(3)
       expect(response.deleted_object_ids).to include(coffee_id)
@@ -445,7 +183,7 @@ describe 'CatalogApi' do
       all_objects = []
 
       while true do
-        response = instance.list_catalog(opts)
+        response = catalog_instance.list_catalog(opts)
         response.objects.each do |obj|
           all_objects.push(obj.id)
         end
@@ -464,7 +202,7 @@ describe 'CatalogApi' do
       opts = {
         include_related_objects: true
       }
-      response = instance.retrieve_catalog_object(object_id, opts)
+      response = catalog_instance.retrieve_catalog_object(object_id, opts)
 
       expect(response.object.id).to eq(object_id)
       expect(response.related_objects.size).to eq(3)
@@ -502,7 +240,7 @@ describe 'CatalogApi' do
           include_deleted_objects: false,
           include_related_objects: false
       }
-      response = instance.search_catalog_objects(body)
+      response = catalog_instance.search_catalog_objects(body)
 
       expect(response.objects.size).to eq(2)
       expect(response.objects[0].type).to eq(SquareConnect::CatalogObjectType::ITEM_VARIATION)
@@ -522,7 +260,7 @@ describe 'CatalogApi' do
         include_deleted_objects: false,
         include_related_objects: false
       }
-      response = instance.search_catalog_objects(body)
+      response = catalog_instance.search_catalog_objects(body)
 
       expect(response.objects.size).to eq(2)
       expect(response.objects[0].type).to eq(SquareConnect::CatalogObjectType::ITEM)
@@ -549,7 +287,7 @@ describe 'CatalogApi' do
       opts = {
         include_related_objects: false
       }
-      object_before = instance.retrieve_catalog_object(coffee_id, opts)
+      object_before = catalog_instance.retrieve_catalog_object(coffee_id, opts)
       expect(object_before.object.item_data.modifier_list_info.size).to eq(1)
       expect(object_before.object.item_data.modifier_list_info[0].modifier_list_id).to eq(milks_id)
 
@@ -557,9 +295,9 @@ describe 'CatalogApi' do
       body.item_ids = [ coffee_id ]
       body.modifier_lists_to_disable = [ milks_id ]
       body.modifier_lists_to_enable = [ syrups_id ]
-      response = instance.update_item_modifier_lists(body)
+      response = catalog_instance.update_item_modifier_lists(body)
 
-      object_after = instance.retrieve_catalog_object(coffee_id, opts)
+      object_after = catalog_instance.retrieve_catalog_object(coffee_id, opts)
       expect(object_after.object.item_data.modifier_list_info.size).to eq(1)
       expect(object_after.object.item_data.modifier_list_info[0].modifier_list_id).to eq(syrups_id)
     end
@@ -570,15 +308,15 @@ describe 'CatalogApi' do
       opts = {
         include_related_objects: false
       }
-      object_before = instance.retrieve_catalog_object(coffee_id, opts)
+      object_before = catalog_instance.retrieve_catalog_object(coffee_id, opts)
       expect(object_before.object.item_data.tax_ids.size).to eq(1)
 
       body = SquareConnect::UpdateItemTaxesRequest.new
       body.item_ids = [ coffee_id ]
       body.taxes_to_disable = [ sales_tax_id ]
-      response = instance.update_item_taxes(body)
+      response = catalog_instance.update_item_taxes(body)
 
-      object_after = instance.retrieve_catalog_object(coffee_id, opts)
+      object_after = catalog_instance.retrieve_catalog_object(coffee_id, opts)
       expect(object_after.object.item_data.tax_ids).to be_nil
     end
   end
@@ -595,7 +333,7 @@ describe 'CatalogApi' do
           percentage: '50.0'
         }
       }
-      response = instance.upsert_catalog_object(body)
+      response = catalog_instance.upsert_catalog_object(body)
 
       expect(response.catalog_object.discount_data.name).to eq("Half off")
       expect(response.catalog_object.id.empty?).to be false
