+++ spec/spec_helper.rb
@@ -64,7 +64,9 @@ RSpec.configure do |config|
       config.access_token = @account["access_token"]
     end
   end
-  
+
+Dir[File.join(__dir__, 'contexts/**/*.rb')].each { |r| require r }
+
 # The settings below are suggested to provide a good initial experience
 # with RSpec, but feel free to customize to your heart's content.
 =begin
